network: moving the firewall validate/parse functions
