+++ azurerm/internal/services/firewall/firewall_policy_resource.go
@@ -7,6 +7,8 @@ import (
 
 	"github.com/hashicorp/go-azure-helpers/response"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/locks"
+	parse2 "github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/firewall/parse"
+	validate2 "github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/firewall/validate"
 
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/location"
 
@@ -16,8 +18,6 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/tf"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/network/parse"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/network/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tags"
 	azSchema "github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tf/schema"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
@@ -34,7 +34,7 @@ func resourceArmFirewallPolicy() *schema.Resource {
 		Delete: resourceArmFirewallPolicyDelete,
 
 		Importer: azSchema.ValidateResourceIDPriorToImport(func(id string) error {
-			_, err := parse.FirewallPolicyID(id)
+			_, err := parse2.FirewallPolicyID(id)
 			return err
 		}),
 
@@ -50,7 +50,7 @@ func resourceArmFirewallPolicy() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: validate.FirewallPolicyName(),
+				ValidateFunc: validate2.FirewallPolicyName(),
 			},
 
 			"resource_group_name": azure.SchemaResourceGroupName(),
@@ -60,7 +60,7 @@ func resourceArmFirewallPolicy() *schema.Resource {
 			"base_policy_id": {
 				Type:         schema.TypeString,
 				Optional:     true,
-				ValidateFunc: validate.FirewallPolicyID,
+				ValidateFunc: validate2.FirewallPolicyID,
 			},
 
 			"dns": {
@@ -221,7 +221,7 @@ func resourceArmFirewallPolicyRead(d *schema.ResourceData, meta interface{}) err
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parse.FirewallPolicyID(d.Id())
+	id, err := parse2.FirewallPolicyID(d.Id())
 	if err != nil {
 		return err
 	}
@@ -279,7 +279,7 @@ func resourceArmFirewallPolicyDelete(d *schema.ResourceData, meta interface{}) e
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parse.FirewallPolicyID(d.Id())
+	id, err := parse2.FirewallPolicyID(d.Id())
 	if err != nil {
 		return err
 	}
