+++ azurerm/internal/services/firewall/validate/firewall_policy_rule_collection_group_name.go
